Rename expandCollapseAll function
