+++ lib/Accordion/ExpandAllButton.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import expandAll from './expandAll';
+import expandAll from './expandCollapseAll';
 import Button from '../Button';
 
 const propTypes = {
