+++ lib/Accordion/expandCollapseAll.js
