+++ lib/Accordion/index.js
@@ -3,4 +3,4 @@ export { default as AccordionSet } from './AccordionSet';
 export { default as DefaultAccordionHeader } from './headers/DefaultAccordionHeader';
 export { default as FilterAccordionHeader } from './headers/FilterAccordionHeader';
 export { default as ExpandAllButton } from './ExpandAllButton';
-export { default as expandAllFunction } from './expandAll';
+export { default as expandAllFunction } from './expandCollapseAll';
